+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Breaking change: `Guard` now takes a `builder` method instead of a `child` property
 * Fixed an issue where subpages wouldn't push on top of a tab page
+* Fixed an issue where query string values were lost after using the back button
 
 # 0.6.2